requirements.txt
# Core Streaming & Vector Store
pathway>=0.1.0
google-generativeai>=0.8.0
tiktoken>=0.7.0

# Agent Framework
langgraph==0.2.45
langchain==0.3.7
langchain-google-genai==2.0.5
langchain-community==0.3.5

# Web & API
fastapi==0.115.5
uvicorn==0.32.1
flask==3.0.3
flask-cors==5.0.0
websockets==13.1
aiohttp==3.10.11

# Document Processing
pypdf==5.1.0
python-docx==1.1.2
beautifulsoup4==4.12.3
lxml==5.3.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
numpy==1.26.4
pandas==2.2.3

# Development
pytest==8.3.4
black==24.10.0
